separate out the ConcurReplicaSVG module as common rendering backend for web.
